pyproject.toml
[project]
name = "learn-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.15",
    "langchain-core>=1.2.27",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.1.6",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
